requirements.txt
google-cloud-translate
google-cloud-firestore
google-cloud-storage
google-generativeai
Pillow
python-dotenv